updates from @mcous's PR comments
