+++ labware-definitions/js/getLabware.js
@@ -2,10 +2,7 @@
 import type {LabwareDefinition} from './types'
 import definitions from '../build/labware.json'
 
-export default function getLabware (labwareName: string): LabwareDefinition {
+export default function getLabware (labwareName: string): ?LabwareDefinition {
   const labware: ?LabwareDefinition = definitions[labwareName]
-  if (!labware) {
-    throw new Error(`No definition found for labware: "${labwareName}"`)
-  }
   return labware
 }
